portfolio/code/legal-case-finder/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
pillow>=10.0.0  # For image apps
fastapi>=0.100.0  # For API extensions
uvicorn>=0.20.0
pydantic>=2.0.0
click>=8.0.0
rich>=13.0.0  # For enhanced console output